# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HST API Documentation",
        "description": "Application API services documentation page",
        "contact": {
            "email": "volkangurgor@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/grades": {
            "get": {
                "tags": [
                    "Grades"
                ],
                "summary": "Get all grades",
                "operationId": "937082441649d3c80b32572bdcd1908b",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Grades"
                ],
                "summary": "Create a new grade",
                "operationId": "09ba8a5b3973d399754f95ab9d83b9ab",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Grade A"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/grades/{id}": {
            "get": {
                "tags": [
                    "Grades"
                ],
                "summary": "Get a specific grade",
                "operationId": "336255077855ed76091fbce88d6d11df",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Grade ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Grade not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Grades"
                ],
                "summary": "Update a grade",
                "operationId": "945be0e2718fba53ecb91c12eca274c1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Grade ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Grade A"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Grade not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Grades"
                ],
                "summary": "Delete a grade",
                "operationId": "f3957d4c8e9fedc3599185c50b5fdc24",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Grade ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Grade not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/classrooms": {
            "get": {
                "tags": [
                    "Classrooms"
                ],
                "summary": "Get all classrooms",
                "operationId": "75528f1f5b94129574dbaacf9379174a",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Classrooms"
                ],
                "summary": "Create a new classroom",
                "operationId": "0714923e4f44a505e6ddafe0f982fd3e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Classroom A"
                                    },
                                    "grade_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "branch_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "Active"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/classrooms/{id}": {
            "get": {
                "tags": [
                    "Classrooms"
                ],
                "summary": "Get a specific classroom",
                "operationId": "fae72f312836772de6d8763a3f47f18c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Classroom ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Classroom not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Classrooms"
                ],
                "summary": "Update a classroom",
                "operationId": "f07fbeb285854acc3bc37f8a71c9447f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Classroom ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Classroom A"
                                    },
                                    "grade_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "branch_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "Active"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Classroom not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Classrooms"
                ],
                "summary": "Delete a classroom",
                "operationId": "256e9910bebaaf34599f5a8d8cd9293c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Classroom ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Classroom not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "in": "header",
                "name": "Authorization",
                "description": "Bearer token to access these API endpoints",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}